+++ packages/expo-updates/android/src/main/java/expo/modules/updates/UpdatesController.kt
@@ -152,23 +152,13 @@ class UpdatesController {
       }
     }
 
-    /**
-     * For [UpdatesModule] to set the [shouldEmitJsEvents] property.
-     */
-    internal var shouldEmitJsEvents: Boolean
-      get() = singletonInstance?.eventManager?.shouldEmitJsEvents ?: false
-      set(value) {
-        singletonInstance?.eventManager?.shouldEmitJsEvents = value
-      }
+    internal fun onEventListenerStartObserving(eventEmitter: EventEmitter?) {
+      singletonInstance?.eventManager?.eventEmitter = eventEmitter
+      singletonInstance?.onEventListenerStartObserving()
+    }
 
-    /**
-     * Binds the [AppContext] and [EventEmitter] instance from [UpdatesModule].
-     */
-    internal fun bindAppContext(appContext: WeakReference<AppContext>, eventEmitter: EventEmitter?) {
-      singletonInstance?.let {
-        it.appContext = appContext
-        it.eventManager.eventEmitter = eventEmitter
-      }
+    internal fun onEventListenerStopObserving() {
+      singletonInstance?.eventManager?.eventEmitter = null
     }
   }
 }